# backend/internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "STA Platform API",
    "version": "v1",
    "description": "Auto-generated shallow spec for the STA Go backend. Method, path, tags and security are accurate; request/response schemas are generic except for a few hand-augmented endpoints. Regenerate with `python3 docs/openapi/generate.py`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "local"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sta_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "adminMFA": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MFA-Code"
      },
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "extraction / agent service token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/api/v1/admin/verification/domains": {
      "get": {
        "operationId": "get_api_v1_admin_verification_domains",
        "summary": "Get admin verification domains",
        "tags": [
          "admin-verification"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_api_v1_admin_verification_domains",
        "summary": "Create admin verification domains",
        "tags": [
          "admin-verification"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verification/domains/{domainID}/active": {
      "post": {
        "operationId": "post_api_v1_admin_verification_domains_domainID_active",
        "summary": "Create admin verification domains domainID active",
        "tags": [
          "admin-verification"
        ],
        "parameters": [
          {
            "name": "domainID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verification/requests/pending": {
      "get": {
        "operationId": "get_api_v1_admin_verification_requests_pending",
        "summary": "Get admin verification requests pending",
        "tags": [
          "admin-verification"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verification/requests/{requestID}/documents": {
      "get": {
        "operationId": "get_api_v1_admin_verification_requests_requestID_documents",
        "summary": "Get admin verification requests requestID documents",
        "tags": [
          "admin-verification"
        ],
        "parameters": [
          {
            "name": "requestID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verification/requests/{requestID}/documents/{documentID}/download": {
      "get": {
        "operationId": "get_api_v1_admin_verification_requests_requestID_documents_documentID_download",
        "summary": "Get admin verification requests requestID documents documentID download",
        "tags": [
          "admin-verification"
        ],
        "parameters": [
          {
            "name": "requestID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verification/requests/{requestID}/review": {
      "post": {
        "operationId": "post_api_v1_admin_verification_requests_requestID_review",
        "summary": "Create admin verification requests requestID review",
        "tags": [
          "admin-verification"
        ],
        "parameters": [
          {
            "name": "requestID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/verification/requests": {
      "get": {
        "operationId": "get_api_v1_verification_requests",
        "summary": "Get verification requests",
        "tags": [
          "verification"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/requests/document": {
      "post": {
        "operationId": "post_api_v1_verification_requests_document",
        "summary": "Create verification requests document",
        "tags": [
          "verification"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/requests/school-email": {
      "post": {
        "operationId": "post_api_v1_verification_requests_school_email",
        "summary": "Create verification requests school email",
        "tags": [
          "verification"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/requests/{requestID}/documents": {
      "post": {
        "operationId": "post_api_v1_verification_requests_requestID_documents",
        "summary": "Create verification requests requestID documents",
        "tags": [
          "verification"
        ],
        "parameters": [
          {
            "name": "requestID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/requests/{requestID}/verify-email": {
      "post": {
        "operationId": "post_api_v1_verification_requests_requestID_verify_email",
        "summary": "Create verification requests requestID verify email",
        "tags": [
          "verification"
        ],
        "parameters": [
          {
            "name": "requestID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}